Move utils to a different location
