+++ src/CreateNewFile/package.json
@@ -3,7 +3,6 @@
   "version": "1.0.0",
   "dependencies": {
     "aws-sdk": "~2",
-    "uuid": "^3.3.2",
-    "utils": "file:../../../src/utils"
+    "uuid": "^3.3.2"
   }
 }
