--- src/utils/index.js
@@ -1,5 +1,0 @@
-const chunkText = require('./chunkText');
-
-module.exports = {
-  chunkText
-};
