--- src/utils/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "utils",
-  "version": "1.0.0",
-  "description": "Utility functions",
-  "main": "index.js"
-}
